requirements.txt
attrs==19.3.0
Automat==0.8.0
Babel==2.9.1
beautifulsoup4==4.10.0
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-ajax-selects==1.9.1
django-allauth==0.40.0
django-appconf==1.0.3
django-autocomplete-light==3.5.1
django-background-tasks==1.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-modal-forms==1.4.4
django-bootstrap4==1.1.1
django-braces==1.13.0
django-ckeditor==6.2.0
django-comments-dab==1.5.0
django-compat==1.0.15
django-cors-headers==3.2.1
django-crispy-forms==1.8.0
django-datatables-view==1.19.0
django-datetimepicker==3.14
django-decouple==2.1
django-dotenv==1.4.2
django-dramatiq==0.9.1
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-formtools==2.2
django-heroku==0.3.1
django-import-export==2.3.0
django-jet==1.0.8
django-js-asset==2.0.0
django-jsonfield==1.4.0
django-mathfilters==1.0.0
django-notify-x==0.1.9
django-otp==0.9.3
django-phonenumber-field==3.0.1
django-restframework==0.0.1
django-select2==7.2.0
django-timezone-field==4.0
django-tinymce==3.4.0
django-two-factor-auth==1.12.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
dramatiq==1.9.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
nexmo==2.4.0
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.9
pexpect==4.6.0
Pillow==9.0.1
prometheus-client==0.13.1
psycopg2==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.18.0
pyserial==3.4
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.3.1
qrcode==6.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sentry-sdk==1.5.8
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.10
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
